refactor: move FlushLogEntriesJob into its own class
